backend/requirements.txt
fastapi==0.115.11
pillow==11.1.0
pymongo==4.11.1
python_jose==3.4.0
python-multipart==0.0.20
torch==2.6.0
transformers==4.49.0
uvicorn==0.34.0
zenml==0.75.0
